Small refactor, generic

+++ gbawj2021_mksgame/include/multi_map_entity.h
@@ -15,7 +15,7 @@ namespace mks
     class MultiMapEntity
     {
     public: 
-        MultiMapEntity(int size);
+        MultiMapEntity(size_t size);
         ~MultiMapEntity();
         
         void init();
