+++ gbawj2021_mksgame/src/multi_ui_element.cpp
@@ -2,11 +2,11 @@
 
 namespace mks
 {
-    MultiUiElement::MultiUiElement(int size)
+    MultiUiElement::MultiUiElement(size_t size)
     {
         active = false;
 
-        for(int index = 0; index < size; index++)
+        for(size_t index = 0; index < size; index++)
         {
             bn::optional<bn::sprite_ptr> new_sprite;
             sprites.push_back(new_sprite);
